# Forge Task/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel-News portal app",
        "version": "1.0.0"
    },
    "paths": {
...
        "/login": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Show the login view",
                "operationId": "d38b0625602ec8e41c3bb98aa2e969de",
                "responses": {
                    "200": {
                        "description": "The login view",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Store the authenticated session",
                "operationId": "42f77e072dec84b0e1094cca1228298e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "The user's email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "description": "The user's password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "remember": {
                                        "description": "Whether to remember the user's session",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect to the intended route",
                        "headers": {
                            "Location": {
                                "description": "The redirect URL",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Destroy the authenticated session",
                "operationId": "03afb12b342c2ea43bdc1de4f39e91a8",
                "responses": {
                    "302": {
                        "description": "Redirect to the home page",
                        "headers": {
                            "Location": {
                                "description": "The redirect URL",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/news/comments": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get a list of comments with news and user details",
                "operationId": "4fd838bdff0f80030a297311b1c8ba12",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserComments"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Create a new comment",
                "operationId": "d06026af1cdd90e34f047c2ede499f64",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "description": "The content of the comment",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "The success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/news/comments/{comment_id}": {
            "put": {
                "tags": [
                    "News"
                ],
                "summary": "Update an existing comment",
                "operationId": "b6cd3164010e96d435195c98b5ccd3ad",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "description": "The ID of the comment to update",
                        "required": true,
                        "schema": {
                            "type": "UUID"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "description": "The new content of the comment",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "The success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Delete a comment by id",
                "operationId": "5cd1e2f142dc6567f24d495a3cfa002a",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "description": "The id of the comment to delete",
                        "required": true,
                        "schema": {
                            "type": "UUID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "The success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "name",
                    "country",
                    "language",
                    "email",
                    "password",
                    "favourite_category"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the user.",
                        "type": "UUID",
                        "format": "string",
                        "readOnly": true
                    },
                    "name": {
                        "description": "The name of the user.",
                        "type": "string"
                    },
                    "country": {
                        "description": "The country of the user.",
                        "type": "string"
                    },
                    "language": {
                        "description": "The language of the user.",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email address of the user.",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "The password of the user.",
                        "type": "string",
                        "format": "password",
                        "writeOnly": true
                    },
                    "favourite_category": {
                        "description": "The favourite category of the user.",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "description": "The date and time when the email address of the user was verified.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "userFavourites": {
                        "description": "The favourites of the user.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserFavourites"
                        },
                        "readOnly": true
                    },
                    "comments": {
                        "description": "The comments of the user.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserComments"
                        },
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "UserComments": {
                "required": [
                    "user_id",
                    "comment"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the user comment.",
                        "type": "UUID",
                        "format": "string",
                        "readOnly": true
                    },
                    "news_id": {
                        "description": "The identifier of the user who made the comment.",
                        "type": "UUID",
                        "format": "string"
                    },
                    "user_id": {
                        "description": "The identifier of the news that has the comments.",
                        "type": "UUID",
                        "format": "string"
                    },
                    "comment": {
                        "description": "The content of the comment.",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "UserFavourites": {
                "required": [
                    "user_id",
                    "title",
                    "description",
                    "newsURL",
                    "imageURL"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the user favourite.",
                        "type": "UUID",
                        "format": "string",
                        "readOnly": true
                    },
                    "user_id": {
                        "description": "The identifier of the user who added the favourite.",
                        "type": "UUID",
                        "format": "string"
                    },
                    "title": {
                        "description": "The title of the favourite.",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the favourite.",
                        "type": "string"
                    },
                    "newsURL": {
                        "description": "The URL of the news article related to the favourite.",
                        "type": "string",
                        "format": "uri"
                    },
                    "imageURL": {
                        "description": "The URL of the image related to the favourite.",
                        "type": "string",
                        "format": "uri"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
        }
    }
}